pyproject.toml
[tool.poetry]
name = "insiders-backend"
version = "1.0.0"
description = ""
authors = ["Michael Brown <mbrown@corcentric.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.68.1"
fastapi-utils = "0.2.1"
FastAPI-SQLAlchemy = "0.2.1"
python-dotenv = "0.15.0"
psycopg2 = "2.8.6"
python-dateutil = "2.8.1"
feedparser = "6.0.8"
requests = "2.25.1"
requests-toolbelt = "0.9.1"
gunicorn = "20.1.0"
pydantic = "1.8.2"
uvicorn = "0.15.0"
email-validator = "1.1.3"
PyJWT = "2.1.0"
itsdangerous = "1.1.0"
cffi = "1.15.0"
bcrypt = "3.2.0"
passlib = "1.7.4"
pycparser = "2.20"
twilio = "7.2.0"
aioredis = "1.3.1"
alembic = "1.7.4"
httpx = "0.20.0"
cryptography = "3.4.8"
pytz = "~=2020.1"
starlette = "~=0.14.2"
aiofiles = "0.7.0"
python-multipart = "0.0.5"
watchgod = "0.7"
websockets = "10.0"
PyYAML = "6.0"
httptools = "0.2.0"
Faker = "9.8.0"
fastapi-cache = "0.1.0"
hiredis = "2.0.0"
fastapi-camelcase = "1.0.3"
sentry-sdk = "1.5.0"
celery = "5.2.2"
SQLAlchemy = "1.4.32"
boto3 = "1.21.4"
botocore = "~=1.24.5"
meilisearch = "~=0.18.1"
sqlalchemy-mixins = "1.5"
py-linq = "1.3.0"
auth0-python = "^3.22.0"
sphinx-autodoc-typehints = "^1.17.0"
sphinx-autoapi = "^1.8.4"
Sphinx = "^4.5.0"
sphinx-rtd-theme = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
